+++ bindings/python/tests/test_read.py
@@ -2,7 +2,6 @@
 
 import numpy as np
 import pytest
-
 import rumi
 from rumi._ffi import _Spec, ffi, lib