+++ bindings/python/tests/test_time.py
@@ -6,7 +6,6 @@ import struct
 
 import numpy as np
 import pytest
-
 import rumi
 
 geozl = pytest.importorskip("geozl")